integrations: Move `get_allowed_bot_types` function.
Move the `get_allowed_bot_types` and required functions
into `user_profile.ts` to avoid circular dependencies.

Preparatory commit for #31389.
