+++ web/src/settings.ts
@@ -23,6 +23,7 @@ import * as settings_sections from "./settings_sections.ts";
 import * as settings_toggle from "./settings_toggle.ts";
 import {current_user, realm} from "./state_data.ts";
 import * as timerender from "./timerender.ts";
+import * as user_profile from "./user_profile.ts";
 import {user_settings} from "./user_settings.ts";
 
 function get_parsed_date_of_joining(): string {
@@ -53,7 +54,7 @@ export function update_lock_icon_in_sidebar(): void {
 
     $(".org-settings-list .locked").show();
 
-    if (settings_bots.can_create_incoming_webhooks()) {
+    if (user_profile.can_create_incoming_webhooks()) {
         $(".org-settings-list li[data-section='bot-list-admin'] .locked").hide();
     }
 
@@ -100,7 +101,7 @@ export function build_page(): void {
         zuliprc: "zuliprc",
         botserverrc: "botserverrc",
         timezones: timezones.timezones,
-        can_create_new_bots: settings_bots.can_create_incoming_webhooks(),
+        can_create_new_bots: user_profile.can_create_incoming_webhooks(),
         settings_label,
         demote_inactive_streams_values: settings_config.demote_inactive_streams_values,
         web_mark_read_on_scroll_policy_values:
@@ -190,7 +191,7 @@ export function initialize(): void {
         is_guest: current_user.is_guest,
         show_uploaded_files_section: realm.max_file_upload_size_mib > 0,
         show_emoji_settings_lock: !settings_data.user_can_add_custom_emoji(),
-        can_create_new_bots: settings_bots.can_create_incoming_webhooks(),
+        can_create_new_bots: user_profile.can_create_incoming_webhooks(),
         can_edit_user_panel:
             current_user.is_admin ||
             settings_data.user_can_create_multiuse_invite() ||
