+++ web/src/settings_bots.ts
@@ -17,9 +17,7 @@ import * as dialog_widget from "./dialog_widget.ts";
 import {$t, $t_html} from "./i18n.ts";
 import * as integration_url_modal from "./integration_url_modal.ts";
 import * as list_widget from "./list_widget.ts";
-import {page_params} from "./page_params.ts";
 import * as people from "./people.ts";
-import * as settings_config from "./settings_config.ts";
 import * as settings_data from "./settings_data.ts";
 import {realm} from "./state_data.ts";
 import type {HTMLSelectOneElement} from "./types.ts";
@@ -43,11 +41,6 @@ type BotInfo = {
     is_active: boolean;
     is_incoming_webhook_bot: boolean;
     zuliprc: string;
-};
-
-type BotType = {
-    type_id: number;
-    name: string;
 };
 
 function add_bot_row(info: BotInfo): void {
@@ -160,43 +153,22 @@ export function generate_botserverrc_content(
     );
 }
 
-export function can_create_new_bots(): boolean {
-    return settings_data.user_has_permission_for_group_setting(
-        realm.realm_can_create_bots_group,
-        "can_create_bots_group",
-        "realm",
-    );
-}
-
-export function can_create_incoming_webhooks(): boolean {
-    // User who have the permission to create any bot can also
-    // create incoming webhooks.
-    return (
-        can_create_new_bots() ||
-        settings_data.user_has_permission_for_group_setting(
-            realm.realm_can_create_write_only_bots_group,
-            "can_create_write_only_bots_group",
-            "realm",
-        )
-    );
-}
-
 export function update_bot_settings_tip($tip_container: JQuery): void {
-    if (can_create_new_bots()) {
+    if (user_profile.can_create_new_bots()) {
         $tip_container.hide();
         return;
     }
 
     const rendered_tip = render_bot_settings_tip({
-        can_create_any_bots: can_create_new_bots(),
-        can_create_incoming_webhooks: can_create_incoming_webhooks(),
+        can_create_any_bots: user_profile.can_create_new_bots(),
+        can_create_incoming_webhooks: user_profile.can_create_incoming_webhooks(),
     });
     $tip_container.show();
     $tip_container.html(rendered_tip);
 }
 
 function update_add_bot_button(): void {
-    if (can_create_incoming_webhooks()) {
+    if (user_profile.can_create_incoming_webhooks()) {
         $("#bot-settings .add-a-new-bot").show();
         $("#admin-bot-list .add-new-bots").show();
         $("#admin-bot-list .manage-your-bots").hide();
@@ -218,27 +190,9 @@ export function update_bot_permissions_ui(): void {
     update_add_bot_button();
 }
 
-export function get_allowed_bot_types(): BotType[] {
-    const allowed_bot_types: BotType[] = [];
-    const bot_types = settings_config.bot_type_values;
-    if (can_create_new_bots()) {
-        allowed_bot_types.push(
-            bot_types.default_bot,
-            bot_types.incoming_webhook_bot,
-            bot_types.outgoing_webhook_bot,
-        );
-        if (page_params.embedded_bots_enabled) {
-            allowed_bot_types.push(bot_types.embedded_bot);
-        }
-    } else if (can_create_incoming_webhooks()) {
-        allowed_bot_types.push(bot_types.incoming_webhook_bot);
-    }
-    return allowed_bot_types;
-}
-
 export function add_a_new_bot(): void {
     const html_body = render_add_new_bot_form({
-        bot_types: get_allowed_bot_types(),
+        bot_types: user_profile.get_allowed_bot_types(),
         realm_embedded_bots: realm.realm_embedded_bots,
         realm_bot_domain: realm.realm_bot_domain,
     });
